re #22
Dodałem serwis obsługujący wszystkie słowniki. Nie ma co się
rozdrabniać, wystarczy jeden na wszystkie.

+++ 84817_84834/project/source/dentalsurgery/src/test/java/pl/poznan/put/dentalsurgery/repository/ToothActivityDictDaoImplTest.java
@@ -71,7 +71,7 @@ public class ToothActivityDictDaoImplTest extends
 		toothActivity = toothActivityDictDao.getToothActivityById(id);
 		toothActivity.setDescription("NewTestDesc");
 		toothActivity.setPrice(150.50);
-		toothActivity.setVat(0);
+		toothActivity.setVat((double) 0);
 		toothActivity.setAllTooth(false);
 
 		toothActivityDictDao.updateToothActivity(toothActivity);
